backend/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-decouple==3.8
Pillow==10.4.0
redis==5.0.1
celery==5.3.4
django-extensions==3.2.3
requests==2.31.0
cryptography==41.0.8
